Refactor screen handling
Each screen is now in a separate XML file instead of a single monolith.
(For now just mainMenu.xml and game.xml) The way screens are loaded in
has changed as well, so that they are all properly initialised before
they are loaded, removing hacks to work around null pointer exceptions.