Fix bug on going to menu and coming back
When exiting to the main menu and coming back to the game, the game
would crash or no characters would be shown. This is because startRoute
was called on a level loading. This however now happens when finishing
a level as we don't want a route to committed to until it is finished
anyway.